Move Solver to experiment package
